+++ core/src/elements/ons-alert-dialog/index.spec.js
@@ -283,3 +283,12 @@ describe('OnsAlertDialogElement', () => {
     });
   });
 });
+
+  describe('autoStyling', () => {
+    it('adds \'material\' modifier on Android', () => {
+      ons.platform.select('android');
+      let e = document.createElement('ons-alert-dialog');
+      expect(e.getAttribute('modifier')).to.equal('material');
+      ons.platform.select('');
+    });
+  });