+++ core/src/elements/ons-back-button.js
@@ -54,9 +54,6 @@ class BackButtonElement extends BaseElement {
   createdCallback() {
     if (!this.hasAttribute('_compiled')) {
       this._compile();
-      ModifierUtil.initModifier(this, scheme);
-
-      this.setAttribute('_compiled', '');
     }
 
     this._options = {};
@@ -64,6 +61,8 @@ class BackButtonElement extends BaseElement {
   }
 
   _compile() {
+    ons._autoStyle.prepare(this);
+
     this.classList.add('back-button');
 
     const label = util.createElement(`
@@ -78,6 +77,10 @@ class BackButtonElement extends BaseElement {
 
     this.appendChild(icon);
     this.appendChild(label);
+
+    ModifierUtil.initModifier(this, scheme);
+
+    this.setAttribute('_compiled', '');
   }
 
   /**