+++ core/src/elements/ons-list.js
@@ -61,8 +61,18 @@ class ListElement extends BaseElement {
    */
 
   createdCallback() {
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+    }
+  }
+
+  _compile() {
+    ons._autoStyle.prepare(this);
+
     this.classList.add('list');
     ModifierUtil.initModifier(this, scheme);
+
+    this.setAttribute('_compiled', '');
   }
 
   attributeChangedCallback(name, last, current) {