+++ core/src/elements/ons-range.js
@@ -56,9 +56,6 @@ class MaterialInputElement extends BaseElement {
   createdCallback() {
     if (!this.hasAttribute('_compiled')) {
       this._compile();
-      ModifierUtil.initModifier(this, scheme);
-
-      this.setAttribute('_compiled', '');
     }
 
     this._updateBoundAttributes();
@@ -66,10 +63,16 @@ class MaterialInputElement extends BaseElement {
   }
 
   _compile() {
+    ons._autoStyle.prepare(this);
+
     this.innerHTML = `
       <input type="range" class="range">
       <div class="range__left"></div>
     `;
+
+    ModifierUtil.initModifier(this, scheme);
+
+    this.setAttribute('_compiled', '');
   }
 
   _onChange() {