+++ core/src/elements/ons-splitter/index.js
@@ -156,11 +156,7 @@ class SplitterElement extends BaseElement {
   _open(side, options = {}) {
     const menu = this._getSideElement(side);
 
-    if (menu) {
-      return menu.open(options);
-    } else {
-      throw new Error('child "ons-splitter-side" element is not found in this element.');
-    }
+    return menu ? menu.open(options) : Promise.reject('child "ons-splitter-side" element is not found in this element.');
   }
 
   /**
@@ -211,11 +207,7 @@ class SplitterElement extends BaseElement {
   _close(side, options = {}) {
     const menu = this._getSideElement(side);
 
-    if (menu) {
-      return menu.close(options);
-    } else {
-      throw new Error('child "ons-splitter-side" element is not found in this element.');
-    }
+    return menu ? menu.close(options) : Promise.reject('child "ons-splitter-side" element is not found in this element.');
   }
 
   /**
@@ -242,11 +234,7 @@ class SplitterElement extends BaseElement {
   _toggle(side, options = {}) {
     const menu = this._getSideElement(side);
 
-    if (menu) {
-      return menu.toggle(options);
-    } else {
-      throw new Error('child "ons-splitter-side" element is not found in this element.');
-    }
+    return menu ? menu.toggle(options) : Promise.reject('child "ons-splitter-side" element is not found in this element.');
   }
 
   /**
@@ -304,11 +292,7 @@ class SplitterElement extends BaseElement {
   loadContentPage(page, options = {}) {
     const content = this._getContentElement();
 
-    if (content) {
-      return content.load(page, options);
-    } else {
-      throw new Error('child "ons-splitter-content" element is not found in this element.');
-    }
+    return content ? content.load(page, options) : Promise.reject('child "ons-splitter-content" element is not found in this element.');
   }
 
   _onDeviceBackButton(handler) {