+++ core/src/elements/ons-switch.spec.js
@@ -194,6 +194,19 @@ describe('OnsSwitchElement', () => {
       expect(element._checkbox.disabled).to.be.false;
       expect(element._checkbox.hasAttribute('disabled')).to.be.false;
     });
+
+    it('changes the inner input ID', () => {
+      element.setAttribute('input-id', 'myID');
+      expect(element._checkbox.id).to.equal('myID');
+    });
   });
 
+  describe('autoStyling', () => {
+    it('adds \'material\' modifier on Android', () => {
+      ons.platform.select('android');
+      let e = document.createElement('ons-switch');
+      expect(e.getAttribute('modifier')).to.equal('material');
+      ons.platform.select('');
+    });
+  });
 });